@@ -6,12 +6,12 @@ import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.TextView;
@@ -19,7 +19,6 @@ import android.widget.TextView;
 import com.xiaoxin.feng.jhang.activity.AccessbilityActivity;
 import com.xiaoxin.feng.jhang.activity.ApkParserActivity;
 import com.xiaoxin.feng.jhang.activity.DouyinPersonActivity;
-import com.xiaoxin.feng.jhang.activity.H5Activity;
 import com.xiaoxin.feng.jhang.activity.MarsTextActivity;
 import com.xiaoxin.feng.jhang.activity.ReadHardActivity;
 import com.xiaoxin.feng.jhang.activity.ReadPicActivity;
@@ -37,7 +36,6 @@ import com.xiaoxin.guid.listfriend.ListFriendActivity;
 import com.xiaoxin.guid.search.disease.SearchDiseaseActivity;
 import com.xiaoxin.guid.search.drug.DrugListActivity;
 import com.xiaoxin.guid.search.hospital.KsSearchHospitalActivity;
-import com.xiaoxin.guid.search.truth.TruthActivity;
 import com.xiaoxin.guid.util.AppUtil;
 import com.xiaoxin.jhang.qqun.QqunActivity;
 
