@@ -1,8 +1,8 @@
 package com.xiaoxin.feng.jhang.activity;
 
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import com.xiaoxin.feng.jhang.R;
 import com.xiaoxin.feng.jhang.databinding.ActivityH5Binding;
 
