@@ -1,6 +1,6 @@
 package com.xiaoxin.feng.jhang.activity;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.xiaoxin.feng.jhang.R;
