@@ -3,7 +3,7 @@ package com.xiaoxin.feng.jhang.activity;
 import android.content.Intent;
 import android.media.ExifInterface;
 import android.net.Uri;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
