@@ -1,7 +1,7 @@
 package com.xiaoxin.feng.jhang.activity;
 
 import android.os.Looper;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.util.Log;
