@@ -3,7 +3,7 @@ package com.xiaoxin.feng.jhang.activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.xiaoxin.feng.jhang.MainActivity;
 import com.xiaoxin.feng.jhang.widget.MyDrawStrategy;
