@@ -2,7 +2,7 @@ package com.xiaoxin.feng.jhang.activity;
 
 import android.content.Intent;
 import android.os.Looper;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
