@@ -1,7 +1,7 @@
 package com.xiaoxin.feng.jhang.activity.animation;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.xiaoxin.feng.jhang.R;
 
