@@ -1,7 +1,7 @@
 package com.xiaoxin.feng.jhang.app;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 
 /**
