@@ -3,8 +3,8 @@ package com.xiaoxin.feng.jhang.util;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Build;
-import android.support.annotation.StringRes;
-import android.support.v4.content.SharedPreferencesCompat;
+import androidx.annotation.StringRes;
+import androidx.core.content.SharedPreferencesCompat;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
