@@ -1,8 +1,7 @@
 package com.xiaoxin.guid;
 
-import android.content.res.AssetManager;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 
