@@ -2,9 +2,9 @@ package com.xiaoxin.guid.activity;
 
 import android.content.Intent;
 import android.database.sqlite.SQLiteDatabase;
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.util.Log;
 
