@@ -1,7 +1,7 @@
 package com.xiaoxin.guid.adapter;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.SparseBooleanArray;
 
 import com.chad.library.adapter.base.BaseQuickAdapter;
