@@ -1,7 +1,7 @@
 package com.xiaoxin.guid.search.disease;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.xiaoxin.guid.R;
 
