@@ -1,7 +1,7 @@
 package com.xiaoxin.guid.search.drug;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.widget.ImageView;
 import android.widget.TextView;
 
