@@ -1,7 +1,7 @@
 package com.xiaoxin.guid.search.hospital;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
