@@ -2,7 +2,7 @@ package com.xiaoxin.guid.search.hospital;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.SparseBooleanArray;
 
 import com.chad.library.adapter.base.BaseQuickAdapter;
