@@ -3,7 +3,7 @@ package com.xiaoxin.library.util;
 import android.annotation.SuppressLint;
 import android.app.Application;
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.lang.reflect.InvocationTargetException;
 
